# Move metadata generation before analysis