--- a/src/test/ui/associated-types/associated-types-ICE-when-projecting-out-of-err.stderr
+++ b/src/test/ui/associated-types/associated-types-ICE-when-projecting-out-of-err.stderr
@@ -1,9 +1,0 @@
-error[E0277]: the trait bound `(): Add<A>` is not satisfied
-  --> $DIR/associated-types-ICE-when-projecting-out-of-err.rs:23:11
-   |
-LL |     r = r + a;
-   |           ^ the trait `Add<A>` is not implemented for `()`
-
-error: aborting due to previous error
-
-For more information about this error, try `rustc --explain E0277`.