… src/test/ui/associated-types/associated-types-ICE-when-projecting-out-of-err.rs
… src/test/ui/associated-types/associated-types-issue-20346.rs
… src/test/ui/associated-types/associated-types-issue-20346.stderr
… src/test/ui/c-variadic/variadic-ffi-4.nll.stderr
… src/test/ui/c-variadic/variadic-ffi-4.rs
… src/test/ui/c-variadic/variadic-ffi-4.stderr