@nnethercote 2019-09-03
The small improvements make sense; analysis takes a lot less time than code generation, so the amount of additional overlap is relatively small.