@nnethercote 2019-09-03
> My reading of the diff is that metadata generation still happens after lowering and more importantly after cfg stripping/macro expansion. Can you confirm?

That is correct. I couldn't push it any earlier because `do_metadata()` requires a `TyCtxt`, which is produced by `Compiler::global_ctxt()`, which relies on the results of `Compiler::lower_to_hir()`, which relies on the results of `Compiler::expansion()`.