@Centril 2019-09-03
Thanks.

The numbers for `syn` still look enticing. cc @dtolnay