@est31 2019-09-03
>  it slightly slows down the common case when an error occurs during analysis, because metadata generation now precedes analysis.

What about doing them in parallel? Of course, it might still be slightly slower on hardware with not enough cores available, but at least there won't be any regression on sufficiently "good" hardware.