@nnethercote 2019-09-04
@alexcrichton said on Zulip:
> rustc has tons of caches internally
> so the time previously spent in analysis may just be shifted to metadata generation now
> (maybe)
> and the analysis phase, which happens later, just hits all the caches and doesn't actually do anything
This may be true. Understanding the ordering of things in rustc is surprisingly hard. I did spend enough time looking at (and refactoring!) `librustc_interface/queries.rs` and `librustc_interface/passes.rs` that I understand `Compiler` quite well, and I got the high-level ordering right. (And diagnostic println statements confirmed this.) It is still possible that there are things going on at a different level that I'm not aware of.

I am still inclined to abandon this effort. I tried what I set out to try, got a result that makes sense (very slight speedups, but much smaller than the original pipelining because analysis is much quicker than codegen), and suggestions for further work are more speculative. If somebody else wants to experiment with it further, please do so.